requirements/base.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-extensions==3.1.1
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
psycopg2==2.8.6
